requirements.txt
hg+https://bitbucket.org/lcdr/pyraknet
peewee>=3.5.0
psycopg2>=2.7.4
bcrypt>=3.1.4
redis>=2.10.6
hiredis>=0.2.0
ujson>=1.35
PyYAML>=3.12